moving script loading out of usemusicplayer
